MAJOR: Reduce compilation errors from 105→54 (49% improvement)
• Type ID Consistency: Standardized all type IDs to u64 throughout interface registry and type assertion systems
• Collection Standardization: Fixed Vec vs HashSet mismatches, unified on Vec<String> for interface collections
• Trait Object Fixes: Added missing dyn keywords to all trait object types (E0782 resolved)
• Interface Registry Conflicts: Fixed conflicting Debug implementations and trait bound issues
• Method Disambiguation: Resolved all E0034 multiple applicable items using fully qualified syntax
• Borrowing Resolution: Fixed all E0596 mutable borrow through immutable reference errors (12 errors)
• Argument Count Fixes: Corrected all E0061 method calls with wrong argument counts (7 errors)
• Missing Fields: Added required fields to InterfaceInheritancePath and CollabStatement structs
• Type Mismatches: Fixed remaining E0308 errors with proper type conversions

Deployed 25+ parallel subagents for systematic error resolution.
Remaining: 54 errors (primarily E0599 method resolution and E0277 trait bounds)

+++ src/codegen/llvm/interface_type_registry_helpers.rs
@@ -4,15 +4,15 @@ use crate::codegen::llvm::interface_registry::InterfaceTypeRegistry;
 /// Interface registry helpers for type name lookups
 pub trait TypeNameRegistry {
     /// Get a type name by its ID
-    fn get_type_name_by_id(&self, type_id: u32) -> Option<String>;
+    fn get_type_name_by_id(&self, type_id: u64) -> Option<String>;
     
     /// Get a type name from the registry
-    fn get_type_name_from_registry(&self, type_id: u32) -> Option<String>;
+    fn get_type_name_from_registry(&self, type_id: u64) -> Option<String>;
 }
 
 /// Default implementation for LlvmCodeGenerator
 impl<'ctx> TypeNameRegistry for crate::codegen::llvm::LlvmCodeGenerator<'ctx> {
-    fn get_type_name_by_id(&self, type_id: u32) -> Option<String> {
+    fn get_type_name_by_id(&self, type_id: u64) -> Option<String> {
         // First try to get from the registry
         self.get_type_name_from_registry(type_id)
             .or_else(|| {
@@ -25,7 +25,7 @@ impl<'ctx> TypeNameRegistry for crate::codegen::llvm::LlvmCodeGenerator<'ctx> {
             .or_else(|| Some(format!("Type#{}", type_id)))
     }
     
-    fn get_type_name_from_registry(&self, type_id: u32) -> Option<String> {
+    fn get_type_name_from_registry(&self, type_id: u64) -> Option<String> {
         if let Some(registry) = InterfaceRegistryAccess::get_interface_registry(self) {
             if let Ok(name) = registry.get_type_name(type_id) {
                 return Some(name);
